Update ZipStreamer.php
If you add an empty directory, the offset isn't recalculated. This offset holds the current position within the local directory. So every directory you will add, writes a local file header which file data are empty. If you don't add the length of the file header, you will lose it.